feat: optimize grant
