+++ AevatarStation.sln
@@ -79,6 +79,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VerifyJsonDesIssue", "sampl
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VerifyStreamApp", "samples\VerifyStreamApp\VerifyStreamApp.csproj", "{64F10B47-10FE-4CF9-A498-817A3F0D2CB5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Aevatar.AuthServer.Grants", "src\Aevatar.AuthServer.Grants\Aevatar.AuthServer.Grants.csproj", "{F68C4098-4B55-4C62-A476-D72864680DA9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Aevatar.AuthServer.Grants.Tests", "test\Aevatar.AuthServer.Grants.Tests\Aevatar.AuthServer.Grants.Tests.csproj", "{E834928B-B7F0-4596-B1FF-AC29A071B881}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -509,6 +513,30 @@ Global
 		{64F10B47-10FE-4CF9-A498-817A3F0D2CB5}.Release|x64.Build.0 = Release|Any CPU
 		{64F10B47-10FE-4CF9-A498-817A3F0D2CB5}.Release|x86.ActiveCfg = Release|Any CPU
 		{64F10B47-10FE-4CF9-A498-817A3F0D2CB5}.Release|x86.Build.0 = Release|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Debug|x64.Build.0 = Debug|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Debug|x86.Build.0 = Debug|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Release|x64.ActiveCfg = Release|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Release|x64.Build.0 = Release|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Release|x86.ActiveCfg = Release|Any CPU
+		{F68C4098-4B55-4C62-A476-D72864680DA9}.Release|x86.Build.0 = Release|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Debug|x64.Build.0 = Debug|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Debug|x86.Build.0 = Debug|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Release|x64.ActiveCfg = Release|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Release|x64.Build.0 = Release|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Release|x86.ActiveCfg = Release|Any CPU
+		{E834928B-B7F0-4596-B1FF-AC29A071B881}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -549,6 +577,8 @@ Global
 		{9367124E-A85D-4067-98C8-42F672DC95AC} = {5D20AA90-6969-D8BD-9DCD-8634F4692FDA}
 		{55E2C369-C501-43DA-A8CA-A6C4D765DD8F} = {5D20AA90-6969-D8BD-9DCD-8634F4692FDA}
 		{64F10B47-10FE-4CF9-A498-817A3F0D2CB5} = {5D20AA90-6969-D8BD-9DCD-8634F4692FDA}
+		{F68C4098-4B55-4C62-A476-D72864680DA9} = {CA9AC87F-097E-4F15-8393-4BC07735A5B0}
+		{E834928B-B7F0-4596-B1FF-AC29A071B881} = {04DBDB01-70F4-4E06-B468-8F87850B22BE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {28315BFD-90E7-4E14-A2EA-F3D23AF4126F}
