+++ src/Aevatar.AuthServer.Grants/Providers/IGithubProvider.cs
@@ -1,0 +1,12 @@
+namespace Aevatar.AuthServer.Grants.Providers;
+
+public interface IGithubProvider
+{
+    Task<GithubUser?> GetUserInfoAsync(string code);
+}
+
+public class GithubUser
+{
+    public long Id { get; set; }
+    public string? Email { get; set; }
+}
