+++ test/Aevatar.AuthServer.Grants.Tests/AevatarApplicationTestBase.cs
@@ -1,0 +1,10 @@
+﻿using Volo.Abp.Modularity;
+
+namespace Aevatar.AuthServer.Grants;
+
+public abstract class AevatarAuthServerGrantsTestBase<TStartupModule> : AevatarTestBase<TStartupModule>
+    where TStartupModule : IAbpModule
+{
+
+}
+
