+++ test/Aevatar.MongoDB.Tests/MongoDb/AuthServerGrants/MongoDBGithubGrantHandlerTests.cs
@@ -1,0 +1,11 @@
+using Aevatar.AuthServer.Grants;
+using Aevatar.MongoDB;
+using Xunit;
+
+namespace Aevatar.MongoDb.AuthServerGrants;
+
+[Collection(AevatarTestConsts.CollectionDefinitionName)]
+public class MongoDBGithubGrantHandlerTests : GithubGrantHandlerTests<AevatarMongoDbTestModule>
+{
+
+}
